chore(test): rename package
